09: InstanceOf float works

+++ InstanceOf.java
@@ -1,7 +1,7 @@
 // 16_777_216
 
 void main() {
-    int i = 16_777_217;
+    float i = 16_777_217;
     if (i instanceof float fl) {
         IO.println("i is an float: " + fl);
     }
